Revert "Productdetail"
