+++ src/components/ProductDetails/AddToCart/AddToCart.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const AddToCart = () => {
+  return <h4>Add To Cart Component</h4>;
+};
+
+export default AddToCart;
