--- src/components/ProductDetails/ProductInfo/AddToCart/QuantityOption.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const QuantityOption = (props) => {
-  return <option quantity={props.quantity}>{props.quantity}</option>;
-};
-
-export default QuantityOption;
