--- src/components/ProductDetails/ProductInfo/AddToCart/SizeOption.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const SizeOption = (props) => {
-  return (
-    <option sku={props.value}>
-      {" "}
-      {props.currentStyle.skus[props.value].size}{" "}
-    </option>
-  );
-};
-
-export default SizeOption;
