+++ src/components/ProductDetails/ProductInfo/ProductCategory.js
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useState, useContext } from "react";
 import { ProductContext } from "../ProductDetails";
 const ProductCategory = () => {
   const product = useContext(ProductContext);
