+++ src/components/ProductDetails/ProductInfo/ProductInfo.js
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useState, useContext } from "react";
 import ProductTitle from "./ProductTitle";
 import ProductRating from "./ProductRating";
 import ProductPrice from "./ProductPrice";
@@ -6,14 +6,12 @@ import ProductOverview from "./ProductOverview";
 import ProductCategory from "./ProductCategory";
 import ShareProduct from "./ShareProduct";
 import StyleSelector from "./StyleSelector/StyleSelector";
-import AddToCart from "./AddToCart/AddToCart";
+import AddToCart from "../AddToCart/AddToCart";
 
 const ProductInfo = () => {
   const infoStyles = {
     borderStyle: "solid",
     borderWidth: "1px",
-    width: "40%",
-    marginLeft: "50%",
   };
 
   return (
