+++ src/components/ProductDetails/ProductInfo/ProductPrice.js
@@ -1,29 +1,9 @@
-import React, { useContext } from "react";
-import { CurrentStyleContext } from "../ProductDetails";
+import React, { useState, useContext } from "react";
+import { ProductContext } from "../ProductDetails";
 const ProductPrice = () => {
-  const { currentStyle } = useContext(CurrentStyleContext);
+  const product = useContext(ProductContext);
 
-  const salePriceStyle = {
-    color: "red",
-  };
-  const oldPriceStyle = {
-    textDecoration: "line-through",
-  };
-
-  return (
-    <h4>
-      Product Price:
-      {currentStyle.sale_price ? (
-        <span>
-          <span style={salePriceStyle}>{currentStyle.sale_price}</span>
-          {" Original Price: "}
-          <span style={oldPriceStyle}>{currentStyle.original_price}</span>
-        </span>
-      ) : (
-        currentStyle.original_price
-      )}
-    </h4>
-  );
+  return <h4>Product Price: {product.default_price}</h4>;
 };
 
 export default ProductPrice;
