+++ src/components/ProductDetails/ProductInfo/ProductRating.js
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useState, useContext } from "react";
 import axios from "../../../apis/atelier";
 import { ReviewsContext } from "../ProductDetails";
