+++ src/components/ProductDetails/ProductInfo/ProductTitle.js
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useState, useContext } from "react";
 import axios from "../../../apis/atelier";
 import { ProductContext } from "../ProductDetails";
