+++ src/components/ProductDetails/ProductInfo/ShareProduct.js
@@ -1,6 +1,8 @@
 import React from "react";
 
 const ShareProduct = () => {
+  // Notes: Get product URL
+
   return <h4>Share on Social Media: [Facebook] [Twitter] [Pinterest]</h4>;
 };
