+++ src/components/ProductDetails/ProductInfo/StyleSelector/StyleSelector.js
@@ -1,39 +1,24 @@
-import React, { useContext, useEffect } from "react";
+import React, { useState, useEffect, useContext } from "react";
 import axios from "../../../../apis/atelier";
 import Style from "./Style";
-import { StylesContext, CurrentStyleContext } from "../../ProductDetails";
-import { CurrentSizeContext, CurrentSkuContext } from "../../ProductDetails";
-import { CurrentQuantityContext } from "../../ProductDetails";
 
 const StyleSelector = () => {
-  const styles = useContext(StylesContext);
-  const { currentStyle, setCurrentStyle } = useContext(CurrentStyleContext);
-  const { currentSize, setCurrentSize } = useContext(CurrentSizeContext);
-  const { currentSku, setCurrentSku } = useContext(CurrentSkuContext);
-  const { setCurrentQuantity } = useContext(CurrentQuantityContext);
+  const [styles, setStyles] = useState([]);
 
-  const styleClickHandler = (style) => {
-    setCurrentStyle(style);
-    setCurrentSize(null);
-    setCurrentSku(null);
-    setCurrentQuantity(null);
-    // console.log('currentStyle:', currentStyle);
-  };
+  useEffect(() => {
+    const fetchData = async () => {
+      const styles = await axios.get(`products/40344/styles`);
+      setStyles(styles.data.results);
+    };
+    fetchData();
+  }, []);
 
   return (
     <h4>
-      Style Selector Component. <br></br>
-      Current Style: {currentStyle.name} <br></br>
+      Style Selector Component.
       <div>
         {styles.map((style) => {
-          return (
-            <Style
-              style={style}
-              currentStyle={currentStyle}
-              key={style.style_id}
-              styleClickHandler={styleClickHandler}
-            />
-          );
+          return <Style style={style} key={style.style_id} />;
         })}
       </div>
     </h4>
